Code Refactor[created te order app and all its model, serializers and view logic from the product_app into the Order_App
