+++ src_codes/core/Order_App/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from Order_App.models import Order
+# Register your models here.
+dmin.site.register(Order)
+
