+++ src_codes/core/Order_App/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class OrderAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'Order_App'
