+++ src_codes/core/products_App/admin.py
@@ -1,8 +1,7 @@
 from django.contrib import admin
-from products_App.models import Products, Order, Category
+from products_App.models import Products, Category
 # Register your models here.
 admin.site.register(Products)
-admin.site.register(Order)
 admin.site.register(Category)
